Split up megafile
