+++ src/constants.js
@@ -1,0 +1,15 @@
+const DEFAULT_QUALITY = 70; // %
+const LQIP_WIDTH = 32; // px
+const MIME_TYPES = {
+  gif: 'image/gif',
+  jpg: 'image/jpeg',
+  png: 'image/png',
+  svg: 'image/svg+xml',
+  webp: 'image/webp',
+};
+
+module.exports = {
+  DEFAULT_QUALITY,
+  LQIP_WIDTH,
+  MIME_TYPES,
+};
